[refactor] ContinuityView를 subview로 분리
